# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HabitFrequency =
  (typeof HabitFrequency)[keyof typeof HabitFrequency];

export const HabitFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface Habit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  color: string;
  frequency: HabitFrequency;
  /** @nullable */
  customDays?: string | null;
  archived: boolean;
  sortOrder: number;
  createdAt: string;
}

export type CreateHabitBodyFrequency =
  (typeof CreateHabitBodyFrequency)[keyof typeof CreateHabitBodyFrequency];

export const CreateHabitBodyFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface CreateHabitBody {
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  color: string;
  frequency: CreateHabitBodyFrequency;
  /** @nullable */
  customDays?: string | null;
}

export type UpdateHabitBodyFrequency =
  (typeof UpdateHabitBodyFrequency)[keyof typeof UpdateHabitBodyFrequency];

export const UpdateHabitBodyFrequency = {
  daily: "daily",
  weekdays: "weekdays",
  weekends: "weekends",
  custom: "custom",
} as const;

export interface UpdateHabitBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  icon?: string;
  color?: string;
  frequency?: UpdateHabitBodyFrequency;
  /** @nullable */
  customDays?: string | null;
  sortOrder?: number;
}

export interface Completion {
  id: number;
  habitId: number;
  completedDate: string;
  createdAt: string;
}

export interface ToggleCompletionBody {
  habitId: number;
  completedDate: string;
}

export interface ToggleCompletionResponse {
  completed: boolean;
  completion?: Completion;
}

export interface DashboardStats {
  totalHabits: number;
  activeHabits: number;
  completionsToday: number;
  completionsThisWeek: number;
  overallCompletionRate: number;
  currentLevel: number;
  totalXp: number;
  xpToNextLevel: number;
  longestStreak: number;
}

export interface HabitStreak {
  habitId: number;
  habitName: string;
  habitColor: string;
  habitIcon: string;
  currentStreak: number;
  bestStreak: number;
}

export interface WeeklyActivity {
  dayOfWeek: string;
  count: number;
}

export type ListCompletionsParams = {
  habitId?: number;
  startDate: string;
  endDate: string;
};