# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all habits
 */
export const ListHabitsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  icon: zod.string(),
  color: zod.string(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]),
  customDays: zod.string().nullish(),
  archived: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListHabitsResponse = zod.array(ListHabitsResponseItem);

/**
 * @summary Create a new habit
 */
export const CreateHabitBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  icon: zod.string(),
  color: zod.string(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]),
  customDays: zod.string().nullish(),
});

/**
 * @summary Get a single habit
 */
export const GetHabitParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHabitResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  icon: zod.string(),
  color: zod.string(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]),
  customDays: zod.string().nullish(),
  archived: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a habit
 */
export const UpdateHabitParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateHabitBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  icon: zod.string().optional(),
  color: zod.string().optional(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]).optional(),
  customDays: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

export const UpdateHabitResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  icon: zod.string(),
  color: zod.string(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]),
  customDays: zod.string().nullish(),
  archived: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a habit
 */
export const DeleteHabitParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle archive status of a habit
 */
export const ToggleArchiveHabitParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleArchiveHabitResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  icon: zod.string(),
  color: zod.string(),
  frequency: zod.enum(["daily", "weekdays", "weekends", "custom"]),
  customDays: zod.string().nullish(),
  archived: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List completions for a date range
 */
export const ListCompletionsQueryParams = zod.object({
  habitId: zod.coerce.number().optional(),
  startDate: zod.date(),
  endDate: zod.date(),
});

export const ListCompletionsResponseItem = zod.object({
  id: zod.number(),
  habitId: zod.number(),
  completedDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const ListCompletionsResponse = zod.array(ListCompletionsResponseItem);

/**
 * @summary Toggle a habit completion for a specific date
 */
export const ToggleCompletionBody = zod.object({
  habitId: zod.number(),
  completedDate: zod.coerce.date(),
});

export const ToggleCompletionResponse = zod.object({
  completed: zod.boolean(),
  completion: zod
    .object({
      id: zod.number(),
      habitId: zod.number(),
      completedDate: zod.coerce.date(),
      createdAt: zod.coerce.date(),
    })
    .optional(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardStatsResponse = zod.object({
  totalHabits: zod.number(),
  activeHabits: zod.number(),
  completionsToday: zod.number(),
  completionsThisWeek: zod.number(),
  overallCompletionRate: zod.number(),
  currentLevel: zod.number(),
  totalXp: zod.number(),
  xpToNextLevel: zod.number(),
  longestStreak: zod.number(),
});

/**
 * @summary Get current and best streaks per habit
 */
export const GetStreaksResponseItem = zod.object({
  habitId: zod.number(),
  habitName: zod.string(),
  habitColor: zod.string(),
  habitIcon: zod.string(),
  currentStreak: zod.number(),
  bestStreak: zod.number(),
});
export const GetStreaksResponse = zod.array(GetStreaksResponseItem);

/**
 * @summary Get completions grouped by day of week for the last 4 weeks
 */
export const GetWeeklyActivityResponseItem = zod.object({
  dayOfWeek: zod.string(),
  count: zod.number(),
});
export const GetWeeklyActivityResponse = zod.array(
  GetWeeklyActivityResponseItem,
);