# lib/api-zod/src/generated/types/createHabitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHabitBodyFrequency } from "./createHabitBodyFrequency";

export interface CreateHabitBody {
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  color: string;
  frequency: CreateHabitBodyFrequency;
  /** @nullable */
  customDays?: string | null;
}