# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalHabits: number;
  activeHabits: number;
  completionsToday: number;
  completionsThisWeek: number;
  overallCompletionRate: number;
  currentLevel: number;
  totalXp: number;
  xpToNextLevel: number;
  longestStreak: number;
}