# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { HabitFrequency } from "./habitFrequency";

export interface Habit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  icon: string;
  color: string;
  frequency: HabitFrequency;
  /** @nullable */
  customDays?: string | null;
  archived: boolean;
  sortOrder: number;
  createdAt: Date;
}