# lib/api-zod/src/generated/types/toggleCompletionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ToggleCompletionBody {
  habitId: number;
  completedDate: Date;
}