# lib/api-zod/src/generated/types/toggleCompletionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Completion } from "./completion";

export interface ToggleCompletionResponse {
  completed: boolean;
  completion?: Completion;
}