# lib/api-zod/src/generated/types/updateHabitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateHabitBodyFrequency } from "./updateHabitBodyFrequency";

export interface UpdateHabitBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  icon?: string;
  color?: string;
  frequency?: UpdateHabitBodyFrequency;
  /** @nullable */
  customDays?: string | null;
  sortOrder?: number;
}