# lib/api-zod/src/generated/types/weeklyActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyActivity {
  dayOfWeek: string;
  count: number;
}